Update the integrative method
In this script, I have updated the four different methods of the integration. The tutorial was mainly used the dataset from various single cell technologies.

To cope with Tuesday and Thursday's work, I have put the COVID datasets to our activities training. Different integration method really took a long time to process the data so that I used this part as the activities.